// src/common/types/enum/resultStatus.ts
export enum ResultStatus {
    Success = 'Success',//200
    Created = 'Created',//201
    NoContent = 'NoContent',//204
    BadRequest = 'BadRequest',//400
    Unauthorized = 'Unauthorized',//401
    Forbidden = 'Forbidden',//403
    NotFound = 'NotFound',//404
    CancelledAction = 'CancelledAction'//500
}